+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/models/ChangeFeedMetaData.java
@@ -1,17 +1,21 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.cosmos.implementation.changefeed.common;
+package com.azure.cosmos.models;
 
 import com.azure.cosmos.implementation.Utils;
+import com.azure.cosmos.util.Beta;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonProcessingException;
+
+import java.time.Instant;
 
 /**
  * Change Feed response meta data
  */
+@Beta(value = Beta.SinceVersion.V4_37_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
 public final class ChangeFeedMetaData {
     @JsonProperty("crts")
-    private long conflictResolutionTimestamp;
+    private Instant conflictResolutionTimestamp;
     @JsonProperty("lsn")
     private long logSequenceNumber;
     @JsonProperty("operationType")
@@ -26,7 +30,8 @@ public final class ChangeFeedMetaData {
      *
      * @return conflict resolution timestamp
      */
-    public long getConflictResolutionTimestamp() {
+    @Beta(value = Beta.SinceVersion.V4_37_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
+    public Instant getConflictResolutionTimestamp() {
         return conflictResolutionTimestamp;
     }
 
@@ -35,6 +40,7 @@ public final class ChangeFeedMetaData {
      *
      * @return current logical sequence number
      */
+    @Beta(value = Beta.SinceVersion.V4_37_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
     public long getLogSequenceNumber() {
         return logSequenceNumber;
     }
@@ -44,6 +50,7 @@ public final class ChangeFeedMetaData {
      *
      * @return change Feed operation type
      */
+    @Beta(value = Beta.SinceVersion.V4_37_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
     public ChangeFeedOperationType getOperationType() {
         return operationType;
     }
@@ -53,6 +60,7 @@ public final class ChangeFeedMetaData {
      *
      * @return previous logical sequence number
      */
+    @Beta(value = Beta.SinceVersion.V4_37_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
     public long getPreviousLogSequenceNumber() {
         return previousLogSequenceNumber;
     }
@@ -63,6 +71,7 @@ public final class ChangeFeedMetaData {
      *
      * @return true if ttlExpiration caused the delete.
      */
+    @Beta(value = Beta.SinceVersion.V4_37_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
     public boolean isTimeToLiveExpired() {
         return timeToLiveExpired;
     }