+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/models/ChangeFeedProcessorItem.java
@@ -1,8 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.cosmos.implementation.changefeed.common;
+package com.azure.cosmos.models;
 
 import com.azure.cosmos.implementation.Utils;
+import com.azure.cosmos.util.Beta;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -14,6 +15,7 @@ import com.fasterxml.jackson.databind.JsonNode;
  *
  * Caller is recommended to type cast {@link JsonNode} to cosmos item structure.
  */
+@Beta(value = Beta.SinceVersion.V4_37_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
 public final class ChangeFeedProcessorItem {
     @JsonProperty("current")
     @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -30,6 +32,7 @@ public final class ChangeFeedProcessorItem {
      *
      * @return change feed current item.
      */
+    @Beta(value = Beta.SinceVersion.V4_37_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
     public JsonNode getCurrent() {
         return current;
     }
@@ -41,6 +44,7 @@ public final class ChangeFeedProcessorItem {
      *
      * @return change feed previous item.
      */
+    @Beta(value = Beta.SinceVersion.V4_37_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
     public JsonNode getPrevious() {
         return previous;
     }
@@ -50,8 +54,22 @@ public final class ChangeFeedProcessorItem {
      *
      * @return change feed metadata.
      */
+    @Beta(value = Beta.SinceVersion.V4_37_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
     public ChangeFeedMetaData getChangeFeedMetaData() {
         return changeFeedMetaData;
+    }
+
+    /**
+     * Helper API to convert this changeFeedProcessorItem instance to raw JsonNode format.
+     *
+     * @return jsonNode format of this changeFeedProcessorItem instance.
+     *
+     * @throws IllegalArgumentException If conversion fails due to incompatible type;
+     * if so, root cause will contain underlying checked exception data binding functionality threw
+     */
+    @Beta(value = Beta.SinceVersion.V4_37_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
+    public JsonNode toJsonNode() {
+        return Utils.getSimpleObjectMapper().convertValue(this, JsonNode.class);
     }
 
     @Override